chore: Move the SLiceList into dedicated namespace
